auth/requirements.txt
Flask==2.2.2
Flask_Login==0.6.2
flask_sqlalchemy==3.0.3
python-dotenv==0.21.1
Werkzeug==2.2.2
